+++ packages/akutil-component-template/src/index.js
@@ -1,5 +1,4 @@
 /** @jsx vdom */
-/* eslint react/no-unknown-property: 0 */
 
 import { define, vdom } from 'skatejs';
 import shadowStyles from './shadow.less';
@@ -9,12 +8,13 @@ const definition = {
     return (
       // JSX requires that there only be a single root element.
       // Incremental DOM doesn't require this.
-      <div class={shadowStyles['class-name']}>
-        // This is required for elements in the shadow root to be styled.
-        // This is wrapped in the <div /> because you can't have more than one
-        // root element.
+      <div>
+        {/* This is required for elements in the shadow root to be styled.
+           This is wrapped in the <div /> because you can't have more than one
+           root element.
+        */}
         <style>{shadowStyles.toString()}</style>
-        <p>I am an {elem.tagName} element!</p>
+        <p className={shadowStyles.locals.myClassName}>I am an {elem.tagName} element!</p>
       </div>
     );
   },