+++ packages/akutil-component-template/CHANGELOG.md
@@ -1,5 +1,12 @@
 <a name="3.0.6"></a>
-## 3.0.6 (2016-07-05)
+## 3.0.6 (2016-07-06)
+
+
+### Bug Fixes
+
+* **jsx:** Comment style ([ff01bb0](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/commits/ff01bb0))
+* **jsx:** pragma ([b9edb0c](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/commits/b9edb0c))
+* **style:** Fixed up shadow style example ([1f2f573](https://aui-team-bot/https://bitbucket.org/atlassian/atlaskit/commits/1f2f573))